requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==11.1.0
qrcode[pil]==8.0
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
resend==2.4.0
Pillow==10.4.0
